todo-agent-mcp/pyproject.toml
[project]
name = "mcp-ollama-sse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.11.3",
    "llama-index>=0.13.2",
    "llama-index-llms-ollama>=0.7.1",
    "llama-index-tools-mcp>=0.4.0",
    "langchain-community>=0.3.10",
    "ollama>=0.5.3",
    "quart>=0.19.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]
